requirements.txt
biopython
requests
tqdm
chromadb
sentence-transformers
openai
pydantic
python-dotenv
tenacity
numpy
scikit-learn